+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FTC16072/OpModes/Driving.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.FTC16072;
+package org.firstinspires.ftc.teamcode.FTC16072.OpModes;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
