+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/FTC16072/OpModes/TestWiring.java
@@ -1,8 +1,9 @@
-package org.firstinspires.ftc.teamcode.FTC16072;
+package org.firstinspires.ftc.teamcode.FTC16072.OpModes;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
+import org.firstinspires.ftc.teamcode.FTC16072.Robot;
 import org.firstinspires.ftc.teamcode.FTC16072.mechanisms.Mechanism;
 import org.firstinspires.ftc.teamcode.FTC16072.tests.QQTest;
